+++ CHANGES.md
@@ -13,6 +13,7 @@ Beta Releases
    * Replaced `color`, `outlineColor`, and `outlineWidth` in `DynamicPath` with a `material` property.
    * CZML property references now use a `#` symbol to separate identifier from property path. `objectId.position` should now be `objectId#position`. 
    * `CesiumWidget.showErrorPanel` now takes a `message` parameter in between the previous `title` and `error` parameters.
+   * Renamed `options.endReferenceFrame` to `options.endTransform` in `CameraFlightPath.createAnimation` and `CameraFlightPath.createAnimationRectangle`.
 * `DynamicObject.id` can now include period characters.
 * `ReferenceProperty` can now handle sub-properties, for example, `myObject#billboard.scale`.
 * Added `Cesium.VERSION` to the combined `Cesium.js` file.